fixed when decline friend request, the user who send can see it

+++ server/controllers/friends.js
@@ -104,10 +104,10 @@ export async function confirmFriend(req, reply) {
       console.log("CoN=>", confirmReq);
       const confirmAccept1 = db
         .prepare(
-          `UPDATE friends SET confirmReq = 1 WHERE (user_id = ? AND friends_id = ?) 
+          `UPDATE friends SET confirmReq = ?, saw = ? WHERE (user_id = ? AND friends_id = ?) 
            OR (user_id = ? AND friends_id = ?)`
         )
-        .run(user_id, friend.id, friend.id, user_id);
+        .run(1,1, user_id, friend.id, friend.id, user_id);
       console.log("CONFIRM =>....", confirmAccept1);
       return reply.code(200).send({ message: "confirmed" });
     }
@@ -141,7 +141,7 @@ export async function declineFriend(req, reply) {
       console.log('CoN=>', confirmReq);
       const declineReq = db
         .prepare(
-          `UPDATE friends SET confirmReq = 0 WHERE friends_id = ? AND user_id = ? `
+          `UPDATE friends SET confirmReq = 0, saw = 1 WHERE friends_id = ? AND user_id = ? `
         )
         .run(user_id, friend.id);
       console.log('DECLINE =>....', declineReq);
@@ -179,7 +179,7 @@ export async function requestFriend(req, reply) {
         u.online
       FROM friends f
       JOIN users u ON u.id = f.user_id
-      WHERE f.friends_id = ? AND f.confirmReq = 0
+      WHERE f.friends_id = ? AND f.confirmReq = 2
     `).all(user_id);
     console.log("YYYY=>", checkRequest)
     
@@ -236,8 +236,8 @@ export async function deleteFriend(req, reply) {
   }
   try {
     const deleteFr = db
-      .prepare(`DELETE FROM friends WHERE user_id = ? AND friends_id = ?`)
-      .run(user_id, friend.id);
+      .prepare(`DELETE FROM friends WHERE (user_id = ? AND friends_id = ?) OR (user_id = ? AND friends_id = ?)`)
+      .run(user_id, friend.id, friend.id, user_id);
     return reply.code(200).send({ deleteFr });
   } catch (err) {
     console.error("Database error:", err.message);
@@ -256,7 +256,7 @@ export async function notificationFriends(req, reply) {
         `SELECT friends.*, users.username
         FROM friends 
         JOIN users ON friends.user_id = users.id 
-        WHERE friends.friends_id = ? AND friends.confirmReq = 2 `
+        WHERE friends.friends_id = ? AND friends.confirmReq = 2  AND saw = 0`
       )
       .all(user_id);
 
@@ -292,7 +292,7 @@ export async function notificationFriends(req, reply) {
         `SELECT friends.*, users.username
         FROM friends 
         JOIN users ON friends.user_id = users.id 
-        WHERE friends.user_id = ? AND friends.confirmReq = 0`
+        WHERE friends.user_id = ? AND friends.confirmReq = 0 AND saw = 1`
       )
       .all(user_id);
 
@@ -320,6 +320,32 @@ export async function notificationFriends(req, reply) {
       seeIfFriendDeclined,
       usernamesDeclined,
     });
+  } catch (err) {
+    console.error('Database error:', err.message);
+    return reply.code(500).send({ message: 'Something went wrong' });
+  }
+}
+
+
+export async function sawAccept(req, reply) {
+  const { user_id, username } = req.body;
+
+  console.log("IIIII=>", user_id)
+  console.log("LLLLLL=>", username)
+
+const username_id = db.prepare(`SELECT id FROM users WHERE username = ?`).get(username)
+
+
+console.log("UUUU=>", username_id.id)
+
+  try {
+    const sawOk = db
+      .prepare(
+        `UPDATE friends SET saw = 2 WHERE user_id = ? AND friends_id = ?`
+      )
+      .run(user_id, username_id.id);
+
+    return reply.code(200).send({ message: 'Saw ok', sawOk });
   } catch (err) {
     console.error('Database error:', err.message);
     return reply.code(500).send({ message: 'Something went wrong' });
